requirements.txt
asgiref==3.10.0
attrs==25.4.0
django==5.2.7
django-filter==25.2
django-guardian==3.2.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pyyaml==6.0.3
referencing==0.37.0
rpds-py==0.28.0
ruff==0.14.2
sqlparse==0.5.3
typing-extensions==4.15.0
uritemplate==4.2.0
PyJWT>=2.9.0
